Processing 3.5.4, finsish ParentGraphicsExtendedRenderer
- Fix bug with Menus not disappering when selecting item
- add Knob class
- Remove dragRelease() again

- ExtendedTextRenderer.draw(): strokeWeight(1) (strike-through width is
now fontSize/15)
- ExtendedTextRenderer: fix percentage lineHeight issue when copying
values for ExtendedTextRenderer

- Update Processing to 3.5.4
- Call resize with window instance size not PApplet.width/height. In
OpenGL mode (P2D/P3D), PApplet knows its new size AFTER Frame
- Also fix some displaying issue and furthermore dont set
no_loopUpdateAgain in Frame.update(), it's wrong!
- Quick solution for animation timing in NO_LOOP mode: just assume 60
fps
- made width/height in Control private: they should be set with
setWidth/Height() usually and if one does not want to call update() by
calling setWidth() there is the method setWidthNoUpdate()
- Minor fixes in Control.java, enhance documentation
- Optimize renderer classes. Make opacity in
ParentGraphicsExtendedRenderer work

+++ guiSET/src/core/Knob.java
@@ -62,7 +62,7 @@ public class Knob extends Slider {
 		pg.fill(getForegroundColor());
 
 		// Draw arc in pie style
-		float angle = (float) (3 * PI / 2) / Math.abs(maxValue - minValue) * (value - minValue);
+		float angle = (3 * PI / 2) / Math.abs(maxValue - minValue) * (value - minValue);
 		pg.noStroke();
 		pg.arc(width * 0.5f, height * 0.5f, valueCircleRadius, valueCircleRadius, 3 * PI / 4, 3 * PI / 4 + angle, PIE);
